Get GHCN-D Station Inventory Information for Multiple Stations

This function retrieves inventory information for specified GHCN-D station(s), including available variables, their coverage periods, and other relevant details.

Usage

get_inventory_aws(station_ids)

Arguments

station_ids

Character vector. One or more 11-character GHCN station identification codes.

Value

A data.table containing inventory information for the specified stations.

Examples

# Get inventory for two stations
get_inventory_aws(c("USW00094728", "USC00305816"))
#>      station_id element
#>          <char>  <char>
#>  1: USC00305816    AWND
#>  2: USC00305816    PRCP
#>  3: USC00305816    SNOW
#>  4: USC00305816    SNWD
#>  5: USC00305816    TMAX
#>  6: USC00305816    TMIN
#>  7: USC00305816    TOBS
#>  8: USC00305816    WT05
#>  9: USW00094728    ACMH
#> 10: USW00094728    ACSH
#> 11: USW00094728    ADPT
#> 12: USW00094728    ASLP
#> 13: USW00094728    ASTP
#> 14: USW00094728    AWBT
#> 15: USW00094728    AWND
#> 16: USW00094728    DAEV
#> 17: USW00094728    DASF
#> 18: USW00094728    DAWM
#> 19: USW00094728    EVAP
#> 20: USW00094728    FMTM
#> 21: USW00094728    MDEV
#> 22: USW00094728    MDSF
#> 23: USW00094728    MDWM
#> 24: USW00094728    PGTM
#> 25: USW00094728    PRCP
#> 26: USW00094728    PSUN
#> 27: USW00094728    RHAV
#> 28: USW00094728    RHMN
#> 29: USW00094728    RHMX
#> 30: USW00094728    SNOW
#> 31: USW00094728    SNWD
#> 32: USW00094728    TAVG
#> 33: USW00094728    TMAX
#> 34: USW00094728    TMIN
#> 35: USW00094728    TOBS
#> 36: USW00094728    TSUN
#> 37: USW00094728    WDF1
#> 38: USW00094728    WDF2
#> 39: USW00094728    WDF5
#> 40: USW00094728    WDFG
#> 41: USW00094728    WDFM
#> 42: USW00094728    WDMV
#> 43: USW00094728    WESD
#> 44: USW00094728    WSF1
#> 45: USW00094728    WSF2
#> 46: USW00094728    WSF5
#> 47: USW00094728    WSFG
#> 48: USW00094728    WSFM
#> 49: USW00094728    WT01
#> 50: USW00094728    WT02
#> 51: USW00094728    WT03
#> 52: USW00094728    WT04
#> 53: USW00094728    WT05
#> 54: USW00094728    WT06
#> 55: USW00094728    WT07
#> 56: USW00094728    WT08
#> 57: USW00094728    WT09
#> 58: USW00094728    WT11
#> 59: USW00094728    WT13
#> 60: USW00094728    WT14
#> 61: USW00094728    WT15
#> 62: USW00094728    WT16
#> 63: USW00094728    WT17
#> 64: USW00094728    WT18
#> 65: USW00094728    WT19
#> 66: USW00094728    WT21
#> 67: USW00094728    WT22
#>      station_id element
#>                                                                       description
#>                                                                            <char>
#>  1:                        Average daily wind speed (tenths of meters per second)
#>  2:                                                  Precipitation (tenths of mm)
#>  3:                                                                 Snowfall (mm)
#>  4:                                                               Snow depth (mm)
#>  5:                               Maximum temperature (tenths of degrees Celsius)
#>  6:                               Minimum temperature (tenths of degrees Celsius)
#>  7:            Temperature at the time of observation (tenths of degrees Celsius)
#>  8:                                                 Hail (may include small hail)
#>  9:                             Average Cloudiness Midnight to Midnight (percent)
#> 10:                                Average Cloudiness Sunrise to Sunset (percent)
#> 11:                                                     Description not available
#> 12:                                                     Description not available
#> 13:                                                     Description not available
#> 14:                                                     Description not available
#> 15:                        Average daily wind speed (tenths of meters per second)
#> 16:                                                     Description not available
#> 17:                                                     Description not available
#> 18:                                                     Description not available
#> 19:                      Evaporation of water from evaporation pan (tenths of mm)
#> 20: Time of Fastest Mile or Fastest 1-minute wind (hours and minutes, i.e., HHMM)
#> 21:                                     Multiday evaporation total (tenths of mm)
#> 22:                                                     Description not available
#> 23:                                                     Description not available
#> 24:                                Peak gust time (hours and minutes, i.e., HHMM)
#> 25:                                                  Precipitation (tenths of mm)
#> 26:                                  Daily percent of possible sunshine (percent)
#> 27:                                                     Description not available
#> 28:                                                     Description not available
#> 29:                                                     Description not available
#> 30:                                                                 Snowfall (mm)
#> 31:                                                               Snow depth (mm)
#> 32:                               Average temperature (tenths of degrees Celsius)
#> 33:                               Maximum temperature (tenths of degrees Celsius)
#> 34:                               Minimum temperature (tenths of degrees Celsius)
#> 35:            Temperature at the time of observation (tenths of degrees Celsius)
#> 36:                                                Daily total sunshine (minutes)
#> 37:                                  Direction of fastest 1-minute wind (degrees)
#> 38:                                  Direction of fastest 2-minute wind (degrees)
#> 39:                                  Direction of fastest 5-second wind (degrees)
#> 40:                                         Direction of peak wind gust (degrees)
#> 41:                                         Fastest mile wind direction (degrees)
#> 42:                                                    24-hour wind movement (km)
#> 43:                         Water equivalent of snow on the ground (tenths of mm)
#> 44:                     Fastest 1-minute wind speed (tenths of meters per second)
#> 45:                     Fastest 2-minute wind speed (tenths of meters per second)
#> 46:                     Fastest 5-second wind speed (tenths of meters per second)
#> 47:                                                    Peak gust wind speed (mph)
#> 48:                         Fastest mile wind speed (tenths of meters per second)
#> 49:                                                 Fog, ice fog, or freezing fog
#> 50:          Heavy fog or heaving freezing fog (not always distinguished from 01)
#> 51:                                                                       Thunder
#> 52:                               Ice pellets, sleet, snow pellets, or small hail
#> 53:                                                 Hail (may include small hail)
#> 54:                                                                 Glaze or rime
#> 55:        Dust, volcanic ash, blowing dust, blowing sand, or blowing obstruction
#> 56:                                                                 Smoke or haze
#> 57:                                                      Blowing or drifting snow
#> 58:                                                        High or damaging winds
#> 59:                                                                          Mist
#> 60:                                                                       Drizzle
#> 61:                                                              Freezing drizzle
#> 62:               Rain (may include freezing rain, drizzle, and freezing drizzle)
#> 63:                                                                 Freezing rain
#> 64:                              Snow, snow pellets, snow grains, or ice crystals
#> 65:                                               Unknown source of precipitation
#> 66:                                                                    Ground fog
#> 67:                                                       Ice fog or freezing fog
#>                                                                       description
#>     first_year last_year coverage_years completeness
#>          <int>     <int>          <num>        <num>
#>  1:       2022      2024              3   1.00000000
#>  2:       1871      2024            154   1.00000000
#>  3:       1884      2024            141   1.00000000
#>  4:       1893      2024            132   1.00000000
#>  5:       1871      2024            154   1.00000000
#>  6:       1871      2024            154   1.00000000
#>  7:       2022      2024              3   1.00000000
#>  8:       1957      1957              1   0.01470588
#>  9:       1965      1990             26   0.43333333
#> 10:       1965      1966              2   0.03333333
#> 11:       2006      2024             19   1.00000000
#> 12:       2006      2024             19   1.00000000
#> 13:       2006      2024             19   1.00000000
#> 14:       2006      2024             19   1.00000000
#> 15:       1984      2024             41   1.00000000
#> 16:       1950      1959             10   0.13333333
#> 17:       1877      1911             35   0.23648649
#> 18:       1956      1959              4   0.05797101
#> 19:       1948      1959             12   0.15584416
#> 20:       1984      2011             28   0.68292683
#> 21:       1950      1959             10   0.13333333
#> 22:       1877      1911             35   0.23648649
#> 23:       1956      1959              4   0.05797101
#> 24:       1966      2017             52   0.88135593
#> 25:       1869      2024            156   1.00000000
#> 26:       1965      1981             17   0.28333333
#> 27:       2006      2024             19   1.00000000
#> 28:       2006      2024             19   1.00000000
#> 29:       2006      2024             19   1.00000000
#> 30:       1869      2024            156   1.00000000
#> 31:       1888      2024            137   1.00000000
#> 32:       1998      2005              8   0.29629630
#> 33:       1869      2024            156   1.00000000
#> 34:       1869      2024            156   1.00000000
#> 35:       1952      1952              1   0.01369863
#> 36:       1965      2021             57   0.95000000
#> 37:       1983      1995             13   0.30952381
#> 38:       1995      2024             30   1.00000000
#> 39:       1995      2024             30   1.00000000
#> 40:       1966      1995             30   0.50847458
#> 41:       1965      1981             17   0.28333333
#> 42:       1948      1959             12   0.15584416
#> 43:       1973      1976              4   0.07692308
#> 44:       1978      1995             18   0.38297872
#> 45:       1995      2024             30   1.00000000
#> 46:       1995      2024             30   1.00000000
#> 47:       1966      1995             30   0.50847458
#> 48:       1965      1981             17   0.28333333
#> 49:       1995      2024             30   1.00000000
#> 50:       1995      2024             30   1.00000000
#> 51:       1962      2024             63   1.00000000
#> 52:       1891      2023            133   0.99253731
#> 53:       1973      2021             49   0.94230769
#> 54:       1963      2024             62   1.00000000
#> 55:       1921      2010             90   0.86538462
#> 56:       1995      2024             30   1.00000000
#> 57:       1892      2010            119   0.89473684
#> 58:       1997      2012             16   0.57142857
#> 59:       1996      2013             18   0.62068966
#> 60:       1984      2013             30   0.73170732
#> 61:       1984      2003             20   0.48780488
#> 62:       1962      2013             52   0.82539683
#> 63:       1984      2011             28   0.68292683
#> 64:       1962      2013             52   0.82539683
#> 65:       1995      2013             19   0.63333333
#> 66:       1999      1999              1   0.03846154
#> 67:       1997      2013             17   0.60714286
#>     first_year last_year coverage_years completeness