GHCN-D Stations Data

A dataset containing information about weather stations in the Global Historical Climatology Network - Daily (GHCN-D).

Usage

ghcn.stations

Format

A data table with 9 variables:

id

character. 11-character station identification code.

latitude

numeric. Latitude of the station in decimal degrees.

longitude

numeric. Longitude of the station in decimal degrees.

elevation

numeric. Elevation of the station in meters.

state

character. U.S. postal code for the state (for U.S. and Canadian stations only).

name

character. Name of the station.

gsn_flag

character. Flag indicating if the station is part of the GCOS Surface Network (GSN).

hcn_crn_flag

character. Flag indicating if the station is part of the U.S. Historical Climatology Network (HCN) or U.S. Climate Reference Network (CRN).

wmo_id

character. World Meteorological Organization (WMO) number for the station.

Source

http://noaa-ghcn-pds.s3.amazonaws.com/ghcnd-stations.txt